Updated UI, refactored code base, confirmation upon save playlist
