+++ .env
@@ -1,0 +1,6 @@
+REACT_APP_CLIENT_ID = "4824b5ae50b14db4b523abf744daed42";
+// REDIRECT_URI = "http://localhost:3000/";
+REACT_APP_REDIRECT_URI = "https://randomify-silk.vercel.app/";
+REACT_APP_AUTH_ENDPOINT = "https://accounts.spotify.com/authorize";
+REACT_APP_RESPONSE_TYPE = "token";
+REACT_APP_SCOPE = "playlist-modify-private playlist-modify-public";
