+++ src/components/Splash.jsx
@@ -3,7 +3,7 @@ import logo from "../imgs/Spotify_Icon.png"
 
 function Splash({ authEndpoint, clientId, redirectUri, responseType, scope }) {
   return (
-    <div className="flex justify-center">
+    <div className="flex justify-center mb-auto">
       <div className="my-16 text-center">
         <div className="text-emerald-500 font-semibold text-base sm:text-3xl md:text-4xl lg:text-5xl">
           <p className="py-3 sm:py-8">In need change of pace?</p>
@@ -20,7 +20,7 @@ function Splash({ authEndpoint, clientId, redirectUri, responseType, scope }) {
             className="border-2 text-xl sm:text-3xl md:text-4xl font-semibold rounded-full border-emerald-500 bg-emerald-600 hover:bg-emerald-500 text-gray-900 my-2 px-5 py-3 sm:py-4"
             href={`${authEndpoint}?client_id=${clientId}&redirect_uri=${redirectUri}&response_type=${responseType}&scope=${scope}`}
           >
-            Login to Spotify <img className="inline w-12 sm:w-16" src={logo} alt="Spotify Icon" />
+            Login to Spotify <img className="inline w-9 ml- sm:w-16" src={logo} alt="Spotify Icon" />
           </a>
         </div>
       </div>
